Use the given coqtop program in the makefile
Also put back coqtop-as-coqc.sh, for the makefile only.

We were previously passing coqtop to coq_makefile as coqc, and that
doesn't work.  There's no way to tell coq_makefile to use coqtop, and if
the user passes --coqc-is-coqtop, they're not required to pass a custom
--coqc, so we can't just use coqc in the makefile.  So we fake it with
coqtop -compile (not -load-vernac-source, because I think that doesn't
produce .glob files).