Kludge around a bug
This function really shouldn't be returning either dicts or tuples, but I'm
currently too drained to figure out which it should return uniformly, so I just
make the method that calls it handle both types.  This is terrible, and should
really be fixed sometime.  I suspect the right thing to do is to always return
a tuple, and move the sorting code into get_imports.