Pass anything in --arg to coq_makefile
This should close #27, as long as `--arg=-impredicative-set` is passed
on the command line.  Arguments are not automatically picked up from
the variables in the file, but I don't expect this to be a big issue,
and the easy fix is to pass anything that is required in `make` along
with `--arg`.