Always insert the prelude require on inline-coqlib
This fixes the bug where [Require Import Coq.Init.Prelude] would be
inserted on the first run, and then the require would be stripped when
inlining other files (because they already require Coq.Init.Prelude),
and then the final inline would fail because Coq.Init.Prelude is no
longer required.  Now we add back the [Require] whenever none of the
things imported by Coq.Init.Prelude are in the list of imports, and also
the list of imports is nonempty.  (If it's empty, then we shouldn't need
to add this, and, in fact, adding it might result in dropping
already-inlined code.)