Also inspect the retcode in minimize-requires
This closes #33.  (Mostly.  I left in the checking of the output,
because I can't recall what my original motivation was for saying that I
should check return codes rather than output.  But it's very easy now to
check only the return code and not the output, if I decide that's
better.)