Fix handling of require inlining
We now maintain the invariant that the files we write always have the
header at the top of them (previously, normalizing requires and adding
the admit header moved the comment header down, and we wrote this file
between passes).  This allows us to refactor the logic a bit.

Alas, this is getting complicated enough that the invariants should
really be documented :-/