Use byte locations from glob files
It turns out that we're already pretty consistent about using text
everywhere, so we don't need to change `find-bug.py`.  It might be nice
to be a bit more disciplined in `split_statements.py`; this would
require thinking more carefully about how newlines are interpreted in
pipes in both Python 2 and Python 3, and how popen communication happens
with bytes vs str in Python 2.

I continue to be pleasantly surprised by how relatively small the
changes required to fix bugs are; I think this is a sign of decent
future-robust engineering.  (I also continue to be unpleasantly
surprised by how involved the changes to the test-suite are; I think
this is a sign of bad design/engineering.)

Fixes #60