Be more robust to utf-8 decoding errors
Fix #176

When .glob files are outdated, we will just fail to split requires.  We
could try to be more fine-grained and gracefully skip over statements
where there's a decoding error, but given how heavily we rely on .glob
files for inlining, I don't think there's much harm in failing to split
`Require`s all-together when the .glob file is outdated.